google-cloud/functions_status_subscriber/requirements.txt
google-cloud-pubsub
requests